uses namespace api for eateries and uses throughout, removes more old routes
